requirements.txt
Keras==2.2.4
sklearn-crfsuite==0.3.6
tensorflow-gpu==1.14.0
tensorflow-hub==0.5.0
git+https://www.github.com/keras-team/keras-contrib.git